Just some changes
Added CSS and started making some functions in javascript. Messed around with the dark mode option that I am working on and will improve once I finish the javascript. Trying to focus on creating and making javascript functions and using the dom. I tried to do an array but I got lost and started by doing something not as complex so I could work my way up. Right now I am going to add a taskbar input. This will display more than one of the tasks I need to do and it will also append a button next to the input that will let you press and delete the input next to it. That will be the next big thing I need to focus on and work on. No more HTML or CSS. I just need to learn and focus on making stuff on javascript for this project and then do all the cool fun stuff later once Im done with JS.