changes#2
Added some more of everything. I spent 2 hours in total just adding and deleting stuff and just nothing worked for me. I have an idea of what I want and need to do. I need to review some CSS and have all the inputs in a row. Make functions for each one with an add button next to them. I'll have to add some if-else statements and that will be a good practice/learning process. I now have a full idea of what I want and what I can do for this project. Messed around with a lot of things and stopped since I didn't know that just javascript which is fine since now I know what I need to learn more of and practice. learning javascript is differently a humble experience so far... Imma just tries and finish this project in the next 3 days and learn more javascript and make more mini projects like this.