changes 3.0
Just been working on javascript dom. I've been spending a lot of time on the dom and trying to figure out connecting online functions. Spend around 3 hours on something that just didn't end up working. I guess I learned some stuff from it which is good. I finally got a base working. Also added some CSS and hover effects with border-radius on the buttons so everything looks good. Going to change the display to have everything mainly in the center. It just seems like everything is on the left-right so I am going to work on that. Also, finishing the dark mode is another thing I want to work on, and going to add javascript to the checkbox and have it do something. Probably got like 5 more hours and going to stop adding stuff so I can move on and practice arrays if else and objects and loops in javascript so I can create a more complex and better working project.